# chore: enable WASM support