--- a/.changeset/nasty-cars-attack.md
+++ b/.changeset/nasty-cars-attack.md
@@ -1,0 +1,6 @@
+---
+'@callstack/reassure-cli': patch
+'reassure': patch
+---
+
+feat: Experimental options to enable WebAssembly (--enable-wasm)