--- a/packages/reassure-cli/src/commands/check-stability.ts
+++ b/packages/reassure-cli/src/commands/check-stability.ts
@@ -14,7 +14,12 @@ export const command: CommandModule<{}, CommonOptions> = {
   command: 'check-stability',
   describe: 'Checks how stable is the current machine by running measurements twice for the same code',
   builder: (yargs) => {
-    return applyCommonOptions(yargs);
+    return applyCommonOptions(yargs).option('enable-wasm', {
+      type: 'boolean',
+      default: false,
+      describe:
+        '(experimental) Enables WebAssembly support in tests by modifying Node flags. This may affect test stability.',
+    });
   },
   handler: (args) => run(args),
 };