<!-- Please provide enough information so that others can review your pull request. -->
<!-- Keep pull requests small and focused on a single change. -->

### Summary

This PR adds support for Reassure `--enable-wasm` option that would replace `--jitless` flag for `--no-opt --no-sparkplug` flags intended to disable Turbo Fan optimizing compiler and Sparkplug non-optimizing compiler in order to maintain execution time stability provided by JIT-less mode.

Resolves #426 
Resolves #421

<!-- What existing problem does the pull request solve? Can you solve the issue with a different approach? -->

### Test plan

<!-- List the steps with which we can test this change. Provide screenshots if this changes anything visual. -->