@andrei-zgirvaci 2023-12-04
@mdjastrzebski thanks for the fix. Indeed, running node with the options: `--no-opt --no-sparkplug` fixes the previous issue I had with MSW.

I also noticed from running `node --v8-options` on node LTS: v20 that the `--opt` option is just an alias for `--turbofan`. Maybe we can replace the `--no-opt` option with `--no-turbofan` to keep it more consistent with the second option: `--no-sparkplug`?

Also, I ran some tests with the `runs` option set to **100**. Here are the following resuls:

1. Took a baseline with the `--jitless` option.
 
2. Ran the same test with the option `--no-turbofan` a few times: ~-10%
```
- <CityWeather /> should render the component without significant performance changes [render]: 0.8 ms → 0.8 ms (-0.1 ms, -9.6%)  | 1 → 1
```

3. Ran the same test with the option `--no-turbofan --no-sparkplug` a few times: ~+5%
```
- <CityWeather /> should render the component without significant performance changes [render]: 0.8 ms → 0.9 ms (+0.0 ms, +4.8%)  | 1 → 1
```

To me it seems that adding `--no-turbofan --no-sparkplug` options keeps the tests pretty close to the `--jitless` option. I also ran the `check-stability` command and everything seemed okay to me!

P.S. I am no expert in reassure so please correct me if I haven't tested this properly!